Code tweaks

+++ src/appleseed/renderer/kernel/lighting/sppm/sppmpasscallback.cpp
@@ -40,6 +40,7 @@
 
 // appleseed.foundation headers.
 #include "foundation/math/hash.h"
+#include "foundation/platform/types.h"
 #include "foundation/utility/job/iabortswitch.h"
 #include "foundation/utility/string.h"
 
@@ -108,7 +109,7 @@ void SPPMPassCallback::on_pass_begin(
     m_stopwatch.start();
 
     // Create a new set of photons.
-    const uint32 pass_hash = mix_uint32(frame.get_noise_seed(), m_pass_number);
+    const uint32 pass_hash = mix_uint32(frame.get_noise_seed(), static_cast<uint32>(m_pass_number));
     m_photons.clear_keep_memory();
     m_photon_tracer.trace_photons(
         m_photons,
